--- src/lib/sui-account-manager/index.ts
@@ -1,2 +1,0 @@
-export { SuiAccountManager } from './sui-account-manager';
-export type { DerivePathParams } from './keypair';
