--- src/lib/sui-rpc-provider/index.ts
@@ -1,2 +1,0 @@
-export { SuiRpcProvider } from './sui-rpc-provider';
-export type { NetworkType } from './default-chain-configs';
