--- src/lib/sui-tx-builder/index.ts
@@ -1,2 +1,0 @@
-export type { SuiTxArg, SuiVecTxArg, SuiObjectArg } from './types';
-export { SuiTxBlock } from './sui-tx-block';
