+++ src/libs/suiAccountManager/types.ts
@@ -1,0 +1,10 @@
+export type AccountMangerParams = {
+  mnemonics?: string;
+  secretKey?: string;
+};
+
+export type DerivePathParams = {
+  accountIndex?: number;
+  isExternal?: boolean;
+  addressIndex?: number;
+};
