+++ src/libs/suiAccountManager/util.ts
@@ -1,4 +1,5 @@
 import { fromB64 } from '@mysten/sui.js';
+
 /**
  * @description This regular expression matches any string that contains only hexadecimal digits (0-9, A-F, a-f).
  * @param str
@@ -17,7 +18,7 @@ export const isBase64 = (str: string) => /^[a-zA-Z0-9+/]+={0,2}$/g.test(str);
  * Convert a hex string to Uint8Array
  * @param hexStr
  */
-export function fromHEX(hexStr: string): Uint8Array {
+export const fromHEX = (hexStr: string): Uint8Array => {
   if (!hexStr) {
     throw new Error('cannot parse empty string to Uint8Array');
   }
@@ -30,7 +31,7 @@ export function fromHEX(hexStr: string): Uint8Array {
     throw new Error(`Unable to parse HEX: ${hexStr}`);
   }
   return Uint8Array.from(intArr);
-}
+};
 
 /**
  * @description Convert a hex or base64 string to Uint8Array
